+++ src/pages/MainPage.jsx
@@ -1,8 +1,8 @@
 import React from "react";
-import Header from "../components/Header";
-import NavBar from "../components/NavBar";
+import NavBar from "../components/NavBar/NavBar";
 import styled from "styled-components";
 import { Outlet } from "react-router-dom";
+import Header from "../components/Header/Header";
 
 const MainPage = () => {
   return (
