+++ src/styles/styled.js
@@ -1,8 +1,8 @@
 // For styled components please ctrl click on the original component to quick find the styling. Styled as according to the creator of styled components' example
-
+// This is the same way Spec
 import { TextField } from "@mui/material";
 import styled from "styled-components";
-import Loading from "../components/Loading";
+import Loading from "../components/Loading/Loading";
 
 export const CenteredLoading = styled(Loading)`
   && {
