+++ src/views/AboutApp/AboutApp.jsx
@@ -1,5 +1,5 @@
 import React from "react";
-import AboutAppAccordian from "../components/AboutAppAccordian";
+import AboutAppAccordian from "../../components/AboutAppAccordian/AboutAppAccordian";
 
 const AboutApp = () => {
   return (
