+++ src/views/Antimedium/Antimedium.jsx
@@ -1,5 +1,5 @@
 import React from "react";
-import { CenteredLoading } from "../styles/styled";
+import { CenteredLoading } from "../../styles/styled";
 
 const Antimedium = () => {
   return (
